requirements.txt
aiodns==1.1.1
aiohttp==1.0.5
async-timeout==1.0.0
cchardet==1.0.0
chardet==2.3.0
colorlog==2.7.0
docker-py==1.10.3
docker-pycreds==0.2.1
envoy==0.0.3
flake8==3.0.4
flake8-double-quotes==0.0.1
flake8-import-order==0.9.2
greenlet==0.4.10
mccabe==0.5.2
motor==0.7b0
msgpack-python==0.4.8
multidict==2.1.2
pep8-naming==0.4.1
pycares==2.1.1
pycodestyle==2.0.0
pyflakes==1.2.3
pymongo==2.9.4
PyYAML==3.12
pyzmq==16.0.0
requests==2.10.0
six==1.10.0
uvloop==0.5.4
websocket-client==0.37.0
zmq==0.0.0
